pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "e-commerceSteamingSimulator"
version = "0.1.0"
description = "A-simulator-for_e-commerce-streaming-data."
authors = [
    { name = "Mbarndouka", email = "mmbarndouka@gmail.com" }
]
requires-python = ">=3.11"

dependencies = [
    "pyspark>=3.5.0",
    "python-dotenv>=1.2.2",
]
[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "flake8>=4.0.0",
    ]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
